Excepciones Personalizadas

+++ ResbarLib/src/Modelo/ManejadorCategorias.java
@@ -18,7 +18,7 @@ public class ManejadorCategorias extends Conexion{
     /* Realiza una petición a la base de datos y devuelve una colección de categorías, si el valor del parametro 
        es TRUE cargara todos los productos que están en esa categoría, si el valor del parametro es FALSE la 
        propiedad “productos” de cada categoría será NULL */
-    public static List<Categoria> obtener(boolean producto){
+    public static List<Categoria> obtener(boolean producto) throws ErrorAplicacion{
         Conexion cn = new Conexion();
         String sql = "SELECT * FROM Categoria";
         ResultSet rs = null;
@@ -47,11 +47,7 @@ public class ManejadorCategorias extends Conexion{
                 categorias.add(c);
             }
         } catch (SQLException | ClassNotFoundException ex) {
-            try {
-                throw new ErrorAplicacion("ManejadorCategorias.Obtener()$Error: " + ex.getMessage());
-            } catch (ErrorAplicacion ex1) {
-                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex1);
-            }
+            throw new ErrorAplicacion("ManejadorCategorias.Obtener()$Error: " + ex.getMessage());
         }finally{
             if (rs != null) {
                 try {
@@ -60,63 +56,83 @@ public class ManejadorCategorias extends Conexion{
                     Logger.getLogger(ManejadorProductos.class.getName()).log(Level.SEVERE, null, ex);
                 }
             }
-            cn.cerrar();
+            try {
+                cn.cerrar();
+            } catch (SQLException ex) {
+                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
         return categorias;
     }  
     
     /* Si se desea modificar el objeto “categoria” este actualizara en la base de datos cuando este ya este 
        modoficado, no se modificara el IdCategoria */
-    public static void actualizar(Categoria c){
-        if(!c.nombre.isEmpty()){
+    public static void actualizar(Categoria c) throws ErrorAplicacion{
+        if(!c.nombre.isEmpty() && c.nombre != null) {
             Conexion cn = new Conexion();
-            String sql = "UPDATE Categoria SET nombre = '"+c.nombre+"' WHERE idCategoria = '"+c.idCategoria+"'";
-            cn.UID(sql);
-            cn.cerrar();
-        }else{
             try {
-                throw new ErrorAplicacion("ManejadorCategorias.actualizar()$El campo nombre no debe estar vacio");
-            } catch (ErrorAplicacion ex) {
-                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex);
+                String sql = "UPDATE Categoria SET nombre = '"+c.nombre+"' WHERE idCategoria = '"+c.idCategoria+"'";
+                cn.UID(sql);
+            } catch (ClassNotFoundException | SQLException ex) {
+               throw new ErrorAplicacion("ManejadorCategorias.actualiazar()$Error: "+ ex.getMessage());
+            }finally{
+                try {
+                    cn.cerrar();
+                } catch (SQLException ex) {
+                    throw new ErrorAplicacion("ManejadorCategorias.actualiazar()$Error: "+ ex.getMessage());
+                }
             }
+        }else{
+            throw new ErrorAplicacion("ManejadorCategorias.actualizar()$El campo nombre no debe estar vacio");
         }
     }
     
      /* Agrega el objeto “categoria” a la base de datos. */
-    public static void insertar(Categoria c){
-        if(c.idCategoria != null && !c.nombre.isEmpty()){
+    public static void insertar(Categoria c) throws ErrorAplicacion{
+        if(c.idCategoria > 0 && (!c.nombre.isEmpty() || c.nombre != null)){
             Conexion cn = new Conexion();
             String sql = "INSERT INTO Categoria(idCategoria, nombre) VALUES('"+c.idCategoria+"', '"+c.nombre+"')";
-            cn.UID(sql);
-            cn.cerrar();
-        }else{
             try {
-                throw new ErrorAplicacion("ManejadorCategorias.insertar()$los campos idCategoria y nombre no pueden ir vacios");
-            } catch (ErrorAplicacion ex) {
-                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex);
+                cn.UID(sql);
+            } catch (ClassNotFoundException | SQLException ex) {
+                throw new ErrorAplicacion("ManejadorCategorias.insertar()$Error: "+ ex.getMessage());
+            }finally{
+                try {
+                    cn.cerrar();
+                } catch (SQLException ex) {
+                    throw new ErrorAplicacion("ManejadorCategorias.insertar()$Error: "+ ex.getMessage());
+                }
             }
+            
+        }else{
+            throw new ErrorAplicacion("ManejadorCategorias.insertar()$los campos idCategoria, nombre son invalidos o estan vacios");
         }
     }
     
     /* Elimina el objeto “categoria” a la base de datos. */
-    public static void eliminar(Categoria c){
-        if(c.idCategoria != null){
+    public static void eliminar(Categoria c) throws ErrorAplicacion{
+        if(c.idCategoria != null && c.idCategoria > 0){
             Conexion cn = new Conexion();
             String sql = "DELETE FROM Categoria WHERE idCategoria = '"+c.idCategoria+"'";
-            cn.UID(sql);
-            cn.cerrar();
-        }else{
             try {
-                throw new ErrorAplicacion("ManejadorCategorias.eliminar()$Campo idCategoria vacio");
-            } catch (ErrorAplicacion ex) {
-                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex);
+                cn.UID(sql);
+            } catch (ClassNotFoundException | SQLException ex) {
+                throw new ErrorAplicacion("ManejadorCategorias.eliminar()$CError: "+ex.getMessage());
+            }finally{
+                try {
+                    cn.cerrar();
+                } catch (SQLException ex) {
+                     throw new ErrorAplicacion("ManejadorCategorias.eliminar()$Error: "+ex.getMessage());
+                }
             }
+        }else{
+            throw new ErrorAplicacion("ManejadorCategorias.eliminar()$Campo idCategoria vacio o es negativo");
         }
     }
     
     /* Obtiene el identificador de categoria, va la base de datos a obtener el ultimo ID de categoria y le suma 
        uno a dicho valor. */
-    public static Integer obtenerId(){
+    public static Integer obtenerId() throws ErrorAplicacion{
         Conexion cn = new Conexion();
         ResultSet rs = null;
         String sql = "SELECT MAX(idCategoria) FROM Categoria";
@@ -127,20 +143,20 @@ public class ManejadorCategorias extends Conexion{
                 id = rs.getInt(1);
             }
         } catch (SQLException | ClassNotFoundException ex) {
-            try {
-                throw new ErrorAplicacion("ManejadorCategorias.obtenerId()$Error: "+ex.getMessage());
-            } catch (ErrorAplicacion ex1) {
-                Logger.getLogger(ManejadorCategorias.class.getName()).log(Level.SEVERE, null, ex1);
-            }
+            throw new ErrorAplicacion("ManejadorCategorias.obtenerId()$Error: "+ex.getMessage());
         }finally{
             if (rs != null) {
                 try {
                     rs.close();
                 } catch (SQLException ex) {
-                    Logger.getLogger(ManejadorProductos.class.getName()).log(Level.SEVERE, null, ex);
+                    throw new ErrorAplicacion("ManejadorCategorias.obtenerId()$Error: "+ex.getMessage());
                 }
             }
-            cn.cerrar();
+            try {
+                cn.cerrar();
+            } catch (SQLException ex) {
+                throw new ErrorAplicacion("ManejadorCategorias.obtenerId()$Error: "+ex.getMessage());
+            }
         }
 
         return id + 1;
